requirements.txt
llama-index-core 
llama-index-readers-docling 
llama-index-node-parser-docling 
llama-index-readers-file 
llama-index-embeddings-ollama 
llama-index-llms-ollama
llama-index-retrievers-bm25
llama-index-vector-stores-postgres
crewai
crewai[tools]
arize-phoenix
fastapi
uvicorn
